requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
python-dotenv==1.0.0
flasgger==0.9.7.1
Flask-Migrate==4.0.5